# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: HaloLight OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events for current user
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"

    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found

    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Leads ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Event:
      type: object
      required: [id, userId, title, eventDate, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        eventDate:
          type: string
          format: date-time
        location:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        status:
          type: string
          enum: [upcoming, active, completed, cancelled]
        notes:
          type: ["string", "null"]
        leadId:
          type: ["string", "null"]
        quoteId:
          type: ["string", "null"]
        contractId:
          type: ["string", "null"]
        invoiceId:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        clientEmail:
          type: ["string", "null"]
        clientPhone:
          type: ["string", "null"]
        clientCompany:
          type: ["string", "null"]
        eventStartTime:
          type: ["string", "null"]
        eventEndTime:
          type: ["string", "null"]
        packageName:
          type: ["string", "null"]
        rentalDuration:
          type: ["string", "null"]
        includedPrints:
          type: ["string", "null"]
        equipmentIds:
          type: ["array", "null"]
          items:
            type: string
        equipmentDescription:
          type: ["string", "null"]
        optionsList:
          type: ["string", "null"]
        revenue:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        paymentStatus:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    EventList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
    EventInput:
      type: object
      required: [title, eventDate]
      properties:
        title:
          type: string
        description:
          type: string
        eventDate:
          type: string
          format: date-time
        location:
          type: string
        type:
          type: string
        notes:
          type: string
        leadId:
          type: ["string", "null"]
        quoteId:
          type: ["string", "null"]
        contractId:
          type: ["string", "null"]
        invoiceId:
          type: ["string", "null"]
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: string
        clientCompany:
          type: string
        eventStartTime:
          type: string
        eventEndTime:
          type: string
        packageName:
          type: string
        rentalDuration:
          type: string
        includedPrints:
          type: string
        equipmentIds:
          type: array
          items:
            type: string
        equipmentDescription:
          type: string
        optionsList:
          type: string
        revenue:
          type: string
        currency:
          type: string
        paymentStatus:
          type: string
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        eventDate:
          type: string
          format: date-time
        location:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [upcoming, active, completed, cancelled]
        notes:
          type: string
        leadId:
          type: ["string", "null"]
        quoteId:
          type: ["string", "null"]
        contractId:
          type: ["string", "null"]
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: string
        clientCompany:
          type: string
        eventStartTime:
          type: string
        eventEndTime:
          type: string
        packageName:
          type: string
        rentalDuration:
          type: string
        includedPrints:
          type: string
        equipmentIds:
          type: array
          items:
            type: string
        equipmentDescription:
          type: string
        optionsList:
          type: string
        revenue:
          type: string
        currency:
          type: string
        paymentStatus:
          type: string
        invoiceId:
          type: ["string", "null"]

    # ─── Leads / CRM ─────────────────────────────────────────────────────────
...